Implement ComputedResourceBlocMixin
Mixin holds the main code for turning normal resource blocs into computed resource blocs

Added classes that include the mixin by default

Mixin depends on AutorunWhileActiveStream, a stream that also reacts to the provided callback like the autorun function, but only while the stream is not paused. Added tests for this class also.