Only override loading state inside MobX reactions
If stream hasn't started yet, listening to the stream will start a load. State should reflect loading as true even though it will take a few event queue ticks later.

However, we shouldn't override state if we're outside a MobX reaction, since this could effect the logic of the bloc. Adjust behavior so overriden state is returned to original value depending on where it's read.